requirements.txt
fastapi
uvicorn[standard]
botbuilder-core
botbuilder-schema